+++ sdk/messaging/eventgrid/aznamespaces/CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ### Features Added
 
-- Added in "binary mode" support, which can improve efficiency when sending a single event (PR#22582)
+- Supports API version `2024-06-01`, including support for release with delay.
+- Added in binary mode support, which can improve efficiency when sending a single event (PR#22582)
 
 ### Breaking Changes
+
+- Sending and receiving opeations have been moved to separate clients (SenderClient and ReceiverClient).
+- Method names have been changed, removing `CloudEvent`. 
+- Topic and subscription name are now set at the Client level, as part of `NewSenderClient` or `NewReceiverClient`.
 
 ### Bugs Fixed